+++ internal/domain/config.go
@@ -1,0 +1,6 @@
+package domain
+
+const AppName = "Scalingo Operator"
+
+// Version is set at build time using ldflags.
+var Version = "dev"